requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
sqlalchemy>=2.0
asyncpg>=0.29
httpx>=0.27
pydantic>=2.6
python-dotenv>=1.0
alembic>=1.13
